jitterutils: Add lockless single reader/writer ringbuffer
The ringbuffer supports one reader and one writer without using
locks. The reader thread only updates the reader index and the writer
only the writer index. Furthermore, the read or write never blocks but
we update a overflow counter to detect this situation.

It is design is based on Andrew Morton's review on a ring
buffer implementation:

"""
A circular buffer implementation needs only head and tail
indices. `size' above appears to be redundant.

Implementation-wise, the head and tail indices should *not* be
constrained to be less than the size of the buffer. They should be
allowed to wrap all the way back to zero. This allows you to
distinguish between the completely-empty and completely-full states
while using 100% of the storage.
"""

The compiler generates a wierd code for ringbuffer_read() do
copy the data out of the ring buffer. We might need to change
that in future but for the time being we can live with it even though
it is not extremely fast.

Link http://lkml.iu.edu/hypermail/linux/kernel/0409.1/2709.html
Signed-off-by: Daniel Wagner <daniel.wagner@siemens.com>