jitterplot: address warnings
In plot_all_cpus() data is a slice of df and to address
the SettingWithCopyWarning, it needs to be identified as
as a copy().  Updated deprecated ymin/ymax with bottom/top

./jitterplot:60: SettingWithCopyWarning:
A value is trying to be set on a copy of a slice from a DataFrame.
Try using .loc[row_indexer,col_indexer] = value instead

The `ymin` argument was deprecated in Matplotlib 3.0 and
will be removed in 3.2. Use `bottom` instead.
  alternative='`bottom`', obj_type='argument')
The `ymax` argument was deprecated in Matplotlib 3.0 and
will be removed in 3.2. Use `top` instead.
  alternative='`top`', obj_type='argument')

Signed-off-by: Sean V Kelley <seanvk.dev@oregontracks.org>