jitterplot: address warnings
In plot_all_cpus() data's assignment of "Time" triggers
the SettingWithCopyWarning. It could be identified as
as a copy(), but that's a waste of a memory spike for a
nonapplicable warning.  So turn off the warning.  Also,
updated the deprecated ymin/ymax with bottom/top

./jitterplot:60: SettingWithCopyWarning:
A value is trying to be set on a copy of a slice from a DataFrame.
Try using .loc[row_indexer,col_indexer] = value instead

The `ymin` argument was deprecated in Matplotlib 3.0 and
will be removed in 3.2. Use `bottom` instead.
 alternative='`bottom`', obj_type='argument')
The `ymax` argument was deprecated in Matplotlib 3.0 and
will be removed in 3.2. Use `top` instead.
  alternative='`top`', obj_type='argument')

Signed-off-by: Sean V Kelley <seanvk.dev@oregontracks.org>
Signed-off-by: Daniel Wagner <wagi@monom.org>