jitterplot: add a way to plot into a file
We might be running in a place where interactive operations are not
possible, let alone graphical interaction based on a popup window.

For those cases add a cmdline argument to plot into a file. When such a
filename is given, we save the figure and continue as before. So one can
still interact and save another manually created copy.

Signed-off-by: Henning Schild <henning.schild@siemens.com>