jitterplot: add option to create CDF plot
A common way to present empirically obtained latencies is a CDF plot.
Here, the x-axis denotes the observed latencies, while the x axis shows
the corresponding percentile. By that, all data (including the
outlayers) can be presented in a single plot.

This patch extends the jitterplot script to support creating such a
CDF plot.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>